# specs/001-crm-intelligent-analytics/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CRM智能分析系统一期接口
  version: 1.2.0
  description: >
    面向一期企业微信 AI 问数、Web 智能分析工作台与智能合同审核模块的接口契约，
    覆盖自然语言问数、企业微信消息接入、查询模板、最近查询、结果查询、
    合同审核任务、审核产物下载、导出控制、治理配置、审计检索与销售日报编排。
servers:
  - url: https://crm-analysis.internal.example.com
    description: 内网生产环境示例地址
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/analysis/queries:
    post:
      tags: [Analysis]
      summary: 提交自然语言分析请求
      operationId: createAnalysisQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisQueryRequest'
      responses:
        '201':
          description: 已创建分析请求或返回补问提示
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisQueryResponse'
        '400':
          description: 请求格式不合法或问题超出当前支持范围
        '403':
          description: 当前用户无权使用分析能力
...
  /api/v1/analysis/queries/{queryId}/exports:
    post:
      tags: [Export]
      summary: 导出指定分析结果
      operationId: createAnalysisExport
      parameters:
        - $ref: '#/components/parameters/QueryId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
      responses:
        '201':
          description: 导出任务已创建并完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '403':
          description: 用户无导出权限或导出超限
        '409':
          description: 分析结果尚不可导出
...
  /api/v1/governance/permissions/overview:
    get:
      tags: [Governance]
      summary: 获取权限中心顶部摘要
      operationId: getAccessGovernanceOverview
      responses:
        '200':
          description: 权限中心摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessGovernanceOverviewResponse'
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/application-super-admin-policy:
    get:
      tags: [Governance]
      summary: 获取应用超级管理员授权策略
      operationId: getApplicationSuperAdminPolicy
      responses:
        '200':
          description: 应用超级管理员授权策略
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSuperAdminPolicyResponse'
        '403':
          description: 当前用户无权访问权限中心
    put:
      tags: [Governance]
      summary: 更新应用超级管理员授权策略
      operationId: updateApplicationSuperAdminPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSuperAdminPolicyRequest'
      responses:
        '200':
          description: 应用超级管理员授权策略更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSuperAdminPolicyResponse'
        '400':
          description: 请求参数不合法
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/analysis-scope-policy:
    get:
      tags: [Governance]
      summary: 获取应用超级管理员授权策略（旧分析全量查询兼容接口）
      operationId: getAnalysisScopePolicy
      deprecated: true
      responses:
        '200':
          description: 兼容返回应用超级管理员授权策略；历史全量查询名单会迁移为用户级授权主体
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSuperAdminPolicyResponse'
        '403':
          description: 当前用户无权访问权限中心
    put:
      tags: [Governance]
      summary: 更新应用超级管理员授权策略（旧分析全量查询兼容接口）
      operationId: updateAnalysisScopePolicy
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ApplicationSuperAdminPolicyRequest'
                - $ref: '#/components/schemas/AnalysisScopePolicyRequest'
      responses:
        '200':
          description: 应用超级管理员授权策略更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSuperAdminPolicyResponse'
        '400':
          description: 请求参数不合法
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/role-permissions:
    get:
      tags: [Governance]
      summary: 获取角色权限矩阵
      operationId: listRolePermissions
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          description: 按角色名称关键词筛选
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
          description: 按角色权限状态筛选
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: 角色权限矩阵
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionListResponse'
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/role-permissions/{roleId}:
    put:
      tags: [Governance]
      summary: 更新指定 CRM 角色的权限矩阵
      operationId: updateRolePermission
      parameters:
        - in: path
          name: roleId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRolePermissionRequest'
      responses:
        '200':
          description: 角色权限已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionResponse'
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/channels/wecom-bot/pilot-policy:
    get:
      tags: [Governance]
      summary: 获取企业微信机器人灰度策略
      operationId: getWecomPilotPolicy
      responses:
        '200':
          description: 当前企业微信灰度策略
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WecomPilotPolicyResponse'
        '403':
          description: 当前用户无权访问权限中心
    put:
      tags: [Governance]
      summary: 更新企业微信机器人灰度策略
      operationId: updateWecomPilotPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWecomPilotPolicyRequest'
      responses:
        '200':
          description: 企业微信灰度策略已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WecomPilotPolicyResponse'
        '403':
          description: 当前用户无权访问权限中心
  /api/v1/governance/access-preview:
    post:
      tags: [Governance]
      summary: 预览指定用户当前命中的最终权限
      operationId: previewAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessPreviewRequest'
      responses:
        '200':
          description: 用户最终权限预览结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessPreviewResponse'
        '403':
          description: 当前用户无权访问权限中心
...
  /api/v1/daily-reports/fragments:
    post:
      tags: [DailyReport]
      summary: 记录或补录销售日报片段
      operationId: recordDailyReportFragment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyReportFragmentCreateRequest'
      responses:
        '201':
          description: 日报片段已写入并返回当前日报快照；22 点自动汇总与次日汇总仍以 CRM 源数据口径为准
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReportRecord'
        '400':
          description: 来源接口失败、片段内容为空或片段类型与来源不匹配
  /api/v1/daily-reports:
    get:
      tags: [DailyReport]
      summary: 查询日报列表
      operationId: listDailyReports
      parameters:
        - in: query
          name: businessDate
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/DailyReportStatus'
        - in: query
          name: requesterId
          schema:
            type: string
      responses:
        '200':
          description: 日报列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyReportRecord'
  /api/v1/daily-reports/{reportId}:
    get:
      tags: [DailyReport]
      summary: 查询单条日报
      operationId: getDailyReport
      parameters:
        - $ref: '#/components/parameters/DailyReportId'
      responses:
        '200':
          description: 日报详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReportRecord'
        '404':
          description: 日报不存在
  /api/v1/daily-reports/{reportId}/confirm:
    post:
      tags: [DailyReport]
      summary: 销售确认日报并正式发送主管
      operationId: confirmDailyReport
      parameters:
        - $ref: '#/components/parameters/DailyReportId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyReportConfirmRequest'
      responses:
        '201':
          description: 日报已确认并正式发送主管；若日报中存在明确协助事项，可并行触发协助升级通知
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReportRecord'
        '400':
          description: 日报尚未完成收口或请求参数无效
        '404':
          description: 日报不存在
  /api/v1/daily-reports/audit:
    get:
      tags: [DailyReport]
      summary: 查询日报审计快照
      operationId: listDailyReportAuditSnapshots
      parameters:
        - in: query
          name: businessDate
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/DailyReportStatus'
      responses:
        '200':
          description: 日报审计快照列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyReportAuditSnapshotItem'
...
  /api/v1/daily-reports/cron/reminders:
    post:
      tags: [DailyReport]
      summary: 执行 22:00 CRM 日报源数据检查与个人确认/催报任务
      operationId: runDailyReportReminderSweep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyReportCronRequest'
      responses:
        '200':
          description: 22 点任务已执行；有源数据时返回自动生成的个人日报，无源数据时仅发送本人催报提醒
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyReportRecord'
  /api/v1/daily-reports/cron/close:
    post:
      tags: [DailyReport]
      summary: 执行日报封账任务
      operationId: runDailyReportClosureSweep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyReportCronRequest'
      responses:
        '200':
          description: 封账任务已执行
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyReportRecord'
  /api/v1/daily-reports/cron/summaries:
    post:
      tags: [DailyReport]
      summary: 执行次日 08:00 负责人日报汇总任务
      operationId: runDailyReportSummarySweep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyReportSummaryCronRequest'
      responses:
        '200':
          description: 团队汇总任务已执行，并按负责人返回日报汇总批次
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReportSummaryBatchRecord'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crm_auth_session
    wecomSignatureAuth:
      type: apiKey
      in: header
      name: X-Wecom-Signature
  parameters:
    QueryId:
      in: path
      name: queryId
      required: true
      schema:
        type: string
      description: 分析请求 ID
...
    DailyReportId:
      in: path
      name: reportId
      required: true
      schema:
        type: string
      description: 销售日报 ID
  schemas:
...
    CreateAnalysisQueryRequest:
      type: object
      required: [querySource, channel]
      properties:
        querySource:
          $ref: '#/components/schemas/QuerySourceType'
        channel:
          $ref: '#/components/schemas/ChannelType'
        questionText:
          type: string
          minLength: 1
          maxLength: 500
        templateId:
          type: string
        historyId:
          type: string
        followUpQueryId:
          type: string
        clarificationAnswer:
          type: object
          additionalProperties:
            type: string
        sessionId:
          type: string
    AnalysisQueryResponse:
      type: object
      required: [queryId, status, createdAt]
      properties:
        queryId:
          type: string
        status:
          $ref: '#/components/schemas/AnalysisRequestStatus'
        clarificationPrompt:
          type: string
        missingConditions:
          type: array
          items:
            type: string
        sessionId:
          type: string
        queueNotice:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateExportRequest:
      type: object
      properties:
        format:
          type: string
          enum: [xlsx, csv]
          default: xlsx
    ExportResponse:
      type: object
      required: [exportId, status, rowCount, createdAt]
      properties:
        exportId:
          type: string
        status:
          $ref: '#/components/schemas/ExportStatus'
        rowCount:
          type: integer
          minimum: 0
        blockedReason:
          type: string
        downloadUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        exportedAt:
          type: string
          format: date-time
...
    AccessGovernanceOverviewResponse:
      type: object
      required:
        - analysisEnabledRoleCount
        - wecomPilotMode
        - wecomPilotWhitelistUserCount
        - exportEnabledRoleCount
        - identityMappingIssueCount
      properties:
        analysisEnabledRoleCount:
          type: integer
          minimum: 0
        wecomPilotMode:
          $ref: '#/components/schemas/WecomPilotMode'
        wecomPilotWhitelistUserCount:
          type: integer
          minimum: 0
        exportEnabledRoleCount:
          type: integer
          minimum: 0
        identityMappingIssueCount:
          type: integer
          minimum: 0
    RolePermissionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RolePermissionResponse'
    RolePermissionResponse:
      type: object
      required:
        - roleId
        - roleNameSnapshot
        - status
        - visibleMenus
        - actionKeys
        - webConsoleEnabled
        - wecomBotEligible
        - exportAllowed
        - templateManageAllowed
        - contractReviewUploadAllowed
        - contractReviewCrossViewAllowed
        - contractReviewCrossDownloadAllowed
        - updatedBy
        - updatedAt
      properties:
        roleId:
          type: string
        roleNameSnapshot:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        visibleMenus:
          type: array
          items:
            type: string
        actionKeys:
          type: array
          items:
            type: string
        webConsoleEnabled:
          type: boolean
        wecomBotEligible:
          type: boolean
        exportAllowed:
          type: boolean
        templateManageAllowed:
          type: boolean
        contractReviewUploadAllowed:
          type: boolean
        contractReviewCrossViewAllowed:
          type: boolean
        contractReviewCrossDownloadAllowed:
          type: boolean
        updatedBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        changeReason:
          type: string
        simplifiedPermissionProfile:
          $ref: '#/components/schemas/SimplifiedPermissionProfile'
    UpdateRolePermissionRequest:
      type: object
      required:
        - roleNameSnapshot
        - status
      properties:
        roleNameSnapshot:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        visibleMenus:
          type: array
          items:
            type: string
        actionKeys:
          type: array
          items:
            type: string
        webConsoleEnabled:
          type: boolean
        wecomBotEligible:
          type: boolean
        exportAllowed:
          type: boolean
        templateManageAllowed:
          type: boolean
        contractReviewUploadAllowed:
          type: boolean
        contractReviewCrossViewAllowed:
          type: boolean
        contractReviewCrossDownloadAllowed:
          type: boolean
        simplifiedPermissionProfile:
          $ref: '#/components/schemas/SimplifiedPermissionProfile'
        changeReason:
          type: string
      description: 优先提交 simplifiedPermissionProfile；旧字段仍用于兼容回滚到旧权限抽屉的载荷。
...
    WecomPilotPolicyResponse:
      type: object
      required:
        - channel
        - mode
        - allowUserIds
        - allowRoleIds
        - allowDepartmentIds
        - denyUserIds
        - updatedBy
        - updatedAt
      properties:
        channel:
          type: string
          enum: [wecom-bot]
        mode:
          $ref: '#/components/schemas/WecomPilotMode'
        allowUserIds:
          type: array
          items:
            type: string
        allowRoleIds:
          type: array
          items:
            type: string
        allowDepartmentIds:
          type: array
          items:
            type: string
        denyUserIds:
          type: array
          items:
            type: string
        note:
          type: string
        updatedBy:
          type: string
        updatedAt:
          type: string
          format: date-time
    UpdateWecomPilotPolicyRequest:
      type: object
      required: [mode, allowUserIds, allowRoleIds, allowDepartmentIds, denyUserIds]
      properties:
        mode:
          $ref: '#/components/schemas/WecomPilotMode'
        allowUserIds:
          type: array
          items:
            type: string
        allowRoleIds:
          type: array
          items:
            type: string
        allowDepartmentIds:
          type: array
          items:
            type: string
        denyUserIds:
          type: array
          items:
            type: string
        note:
          type: string
    AccessPreviewRequest:
      type: object
      properties:
        crmUserId:
          type: string
        wecomUserId:
          type: string
...
    AccessPreviewResponse:
      type: object
      required:
        - mappingStatus
        - analysisScopeMode
        - analysisScopeSummary
        - roleNames
        - visibleMenus
        - actionKeys
        - scopeSummary
        - wecomBotAccessState
        - contractPermissions
      properties:
        crmUserId:
          type: string
        crmUserName:
          type: string
        wecomUserId:
          type: string
        mappingStatus:
          type: string
          enum: [MAPPED, UNMAPPED, CONFLICTED]
        analysisScopeMode:
          type: string
          enum: [FULL_ANALYSIS_SCOPE, DEPARTMENT_ANALYSIS_SCOPE]
        analysisScopeSummary:
          type: string
        roleNames:
          type: array
          items:
            type: string
        visibleMenus:
          type: array
          items:
            type: string
        actionKeys:
          type: array
          items:
            type: string
        scopeSummary:
          type: string
        wecomBotAccessState:
          $ref: '#/components/schemas/WecomBotAccessState'
        wecomBotAccessReason:
          type: string
        contractPermissions:
          $ref: '#/components/schemas/ContractPermissionSnapshot'
        simplifiedPermissionProfile:
          $ref: '#/components/schemas/SimplifiedPermissionProfile'
...
    ApplicationSuperAdminPolicyRequest:
      type: object
      required: [subjects, changeReason]
      properties:
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationSuperAdminSubject'
        changeReason:
          type: string
          maxLength: 200
          description: 中文变更原因，保存授权时必填
    ApplicationSuperAdminPolicyResponse:
      type: object
      required: [policyId, subjects, fullAccessUserIds, fullAccessRoleIds, updatedBy, updatedAt]
      properties:
        policyId:
          type: string
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationSuperAdminSubject'
        fullAccessUserIds:
          type: array
          description: 兼容旧前端的有效用户主体列表
          items:
            type: string
        fullAccessRoleIds:
          type: array
          description: 有效角色主体列表
          items:
            type: string
        updatedBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        changeReason:
          type: string
    AnalysisScopePolicyRequest:
      type: object
      description: 旧分析全量查询授权请求，仅作为应用超级管理员授权兼容格式保留
      required: [fullAccessUserIds, changeReason]
      properties:
        fullAccessUserIds:
          type: array
          items:
            type: string
        changeReason:
          type: string
          maxLength: 200
...
    DailyReportFragmentCreateRequest:
      type: object
      required: [fragmentType, content]
      properties:
        fragmentType:
          $ref: '#/components/schemas/DailyReportFragmentType'
        content:
          type: string
          minLength: 1
        businessDate:
          type: string
          format: date
        supervisorId:
          type: string
          description: 首次生成日报时必填，后续补片段可沿用已有主管
        supervisorName:
          type: string
        sourceLabel:
          type: string
        sourceInterface:
          $ref: '#/components/schemas/DailyReportSourceInterface'
        sourceObjectId:
          type: string
          description: 来源对象 ID，接入上游接口时建议填写
        sourceOperatorId:
          type: string
        sourceOperatorName:
          type: string
        sourceCode:
          type: integer
          description: 上游返回码，只有 0 才允许写入日报
        capturedAt:
          type: string
          format: date-time
    DailyReportConfirmRequest:
      type: object
      properties:
        confirmedAt:
          type: string
          format: date-time
    DailyReportCronRequest:
      type: object
      required: [businessDate]
      properties:
        businessDate:
          type: string
          format: date
        sentAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    DailyReportSummaryCronRequest:
      type: object
      required: [businessDate]
      properties:
        businessDate:
          type: string
          format: date
        recipientIds:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
...
    DailyReportStatus:
      type: string
      enum: [DRAFT, PENDING_CONFIRMATION, CONFIRMED, LATE_CONFIRMED, CLOSED, SUMMARIZED]
...
    QuerySourceType:
      type: string
      enum: [FREE_TEXT, COMMON_TEMPLATE, RECENT_RERUN]
    ChannelType:
      type: string
      enum: [wecom-bot, web-console]
...
    ExportStatus:
      type: string
      enum: [REQUESTED, COMPLETED, BLOCKED]
...